Added swipe to refresh and made more cleaner
