+++ android/app/src/main/AndroidManifest.xml
@@ -7,7 +7,8 @@
          additional functionality it is fine to subclass or reimplement
          FlutterApplication and put your custom class here. -->
     <application
-        android:name="io.flutter.app.FlutterApplication"
+        android:name="${applicationName}"
+        android:exported="true"
         android:label="News"
         android:icon="@mipmap/ic_launcher">
         <activity
