+++ lib/presentation/widgets/loading_widget.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class LoadingWidget extends StatelessWidget {
+  const LoadingWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const CircularProgressIndicator(
+      backgroundColor: Colors.yellow,
+    );
+  }
+}
